+++ config/guildlevels.rb
@@ -1,10 +1,14 @@
 module DistoolsGuilds
   # Guilds (friendguilds, supportguild)
-  # Looking for donor guilds? It's in Mongo.
-  # friendGuilds, array (int). additional perms linked to friendGuilds.
-  friendGuilds = []
-  # supportGuild, int. the support guild
-  supportGuild = 
-  # testGuild, int. guild where config/semi-sensitive options are outputted.
-  testGuild = 
+  # FRIENDGUILDS, array (int). additional perms linked to friendGuilds.
+  FRIENDGUILDS = [1166235874208518216,1142527103187812383]
+  # SUPPORTGUILD, int. the support guild
+  SUPPORTGUILD = 
+  # TESTGUILD, int. guild where config/semi-sensitive options are outputted.
+  TESTGUILD = 
+  # BETAGUILDS, array (int). guilds where beta commands are enabled.
+  BETAGUILDS = []
+  # BLACKLISTEDGUILDS, array (int). guilds where distools is disabled.
+  # ** PUT IN GUILD IDS, NOT USER IDS **
+  BLACKLISTEDGUILDS = []
 end