+++ main.rb
@@ -22,8 +22,9 @@ require "mongo"
 require_relative "config/perms"
 require_relative "config/mongo"
 require_relative "config/strings"
-database = Mongo::Client.new(DataConn::URI)
-puts "(Maybe) connected to Mongo."
+require_relative "config/eval"
+require_relative "config/prefs"
+require_relative "config/guildlevels"
 log = Log4r::Logger.new("discorb")
 log.add Log4r::FileOutputter.new("filelog",
                                  filename: "distools.log",
@@ -33,17 +34,29 @@ log.add Log4r::IOOutputter.new("stderr", $stderr,
                                level: Logger::WARN)
 systemUsername = Etc.getpwuid(Process.uid).name
 uname = Etc.uname
-friendGuilds = [961703572741951518,972546971460059197]
-botVersion = "2.1 RW"
+botVersion = "2.1.0 RW"
 commit = `git rev-parse --short HEAD`
 commitMsg = `git show-branch --no-name`
 
 File.foreach("index.txt") { |line| puts line }
 puts "loaded bot owners: #{BotPerms::BOTOWNERS}"
 puts "loaded blacklist: #{BotPerms::BLACKLISTED}"
-puts "loaded friend guilds: #{friendGuilds}"
 Dotenv.load  # Loads .env file
 puts "loaded .env and thus i am proceeding"
+puts "!! ATTENTION !!"
+puts "A breaking change might be coming up soon (idk though): Message intents will be disabled with the conversion of most text-based commands to slash commands."
+puts "However, SUPERSPY requires intents (due to the fact it logs messages), so if you have SUPERSPY enabled, this message does not apply."
+puts "Bot startup will continue in 5 seconds."
+sleep 1
+puts "Bot startup will continue in 4 seconds."
+sleep 1
+puts "Bot startup will continue in 3 seconds."
+sleep 1
+puts "Bot startup will continue in 2 seconds."
+sleep 1
+puts "Bot startup will continue in 1 second."
+sleep 1
+puts "Continuing startup."
 intents = Discorb::Intents.new(members: true, message_content: true, presences: true)
 client = Discorb::Client.new(logger: log, intents: intents)  # Create client for connecting to Discord
 # activity = Discorb::Activity.new Activiteh!!!
@@ -53,7 +66,7 @@ client = Discorb::Client.new(logger: log, intents: intents)  # Create client for
   client.on event do
     client.update_presence(
       Discorb::Activity.new(
-        "#{client.guilds.length} guilds, Ruby #{RUBY_VERSION} // Discorb #{Discorb::VERSION}", :watching
+        "#{client.guilds.length} guilds on Ruby #{RUBY_VERSION}", :watching
       ),
       status: :dnd
     )
@@ -71,7 +84,10 @@ client.slash("guild", "returns guild information") do |interaction|
 # Before absolutely anything run the Blacklist Test
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+ elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
+  # Looks good. Now we can run the command.
   if friendGuilds.include?(interaction.guild.id) === true
 	if interaction.guild.description.nil? === false
 	interaction.post(embed: Discorb::Embed.new("Guild information", "Guild name: #{interaction.guild.name}\nThis server is a friend guild (in the `friendGuilds` array)\nFeatures: `#{interaction.guild.features}`\nGuild ID: `#{interaction.guild.id}`\nGuild description:\n```\n#{interaction.guild.description}\n```\nMember count: #{interaction.guild.member_count}\nBot joined (UTC): `#{interaction.guild.joined_at}`", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
@@ -93,6 +109,8 @@ end
 client.slash("invite", "gives you 2 invite links: limited and full") do |interaction|
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+ elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("Invite me!", "I have two invite links. One is the limited version and one is full admin permissions.\n\n[Limited](https://discord.com/api/oauth2/authorize?client_id=#{ENV['ID']}&permissions=1634771397751&scope=applications.commands%20bot)\n[Full](https://discord.com/api/oauth2/authorize?client_id=#{ENV['ID']}&permissions=8&scope=applications.commands%20bot)", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 end
@@ -101,7 +119,9 @@ end
 # best interaction ever
 client.slash("garment", "garment") do |interaction|
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
-	interaction.post("this is a certified not garment moment", embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)))
+  interaction.post("this is a certified not garment moment", embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)))
+ elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
 	interaction.post("garment", embed: Discorb::Embed.new("garment", "garment", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  end
@@ -109,6 +129,8 @@ end
 client.slash("instance", "Details about this instance of distools including owners, blacklist, and more") do |interaction|
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+ elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("About this instance", "The description of this instance is: #{DistoolsStr::INSTDESCRIPTION}\ndistools version: #{botVersion}\n#{client.user} is on distools commit #{commit} with the commit message:\n```\n#{commitMsg}```", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 end
@@ -117,6 +139,8 @@ end
 client.slash("hello", "bot stats") do |interaction|
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+ elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("distools #{botVersion} | hello world :)", "I am Distools, a bot made by Winfinity\#1252.", color: Discorb::Color.from_rgb(201, 0, 0), fields: [Discorb::Embed::Field.new("ruby ver", "`#{RUBY_VERSION}-p#{RUBY_PATCHLEVEL}`", inline: false), Discorb::Embed::Field.new("discorb ver", "`#{Discorb::VERSION}`", inline: false), Discorb::Embed::Field.new("OS", "`#{uname[:sysname]} #{uname[:release]}`", inline: false), Discorb::Embed::Field.new("running as user", "`#{systemUsername}`", inline: false)]), ephemeral: false)
 end
@@ -125,12 +149,14 @@ end
 client.slash("help", "gives you my commands") do |interaction|
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
   interaction.post(embed: Discorb::Embed.new("Help | distools #{botVersion}", "I am Distools. A bot made with :heart: (+ caffiene and Ruby) by Winfinity#1252.\n\n**Command List**\n`/hello` - (Interaction) Gives you a list of bot stats.\n`/instance` - (Interaction) Gives you information about this distools instance.\n`/help` - (Interaction) Lists commands.\n`/stop` - (Interaction, bot owners only) Stops the bot using `exit`.\n`/guild` - (Interaction) Guild information.\n`/whoami` - (Interaction) You are #{interaction.user.name}. Seriously, this gives you more information.\n`/why-god-why` - (Interaction, bot owners only) Tests everything. Template command.\n`dist.eval` - (Bot owners only) Evaluates a piece of ruby code.\n`dist.exec` - (\"Super\" bot owners only) Executes a script on the machine distools is running on (via the backticks)\n`/am-i-a-dev` - (Interaction) Tells you whether you are listed as a dev or not.\n`/invite` - (Interaction) Returns invite links.\n\nDistools is open source (the code is [right here](https://github.com/haydenwalker980)) - meaning you are free to use, modify, and distribute it. If you enjoy distools and want to keep the lights on when we inevitably move to a server, [please consider contributing](https://liberapay.com/Win) via liberapay. To have your donation count against an indicator in user information, leave your Discord ID in your liberapay description.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 end
 end
 
-client.slash("eval", "evaluates ruby code", {
+client.slash("eval", "evaluates ruby code (owner-only)", {
   "code" => { 
     required: true,
     type: :string,
@@ -195,7 +221,7 @@ end
 client.on :message do |message|
   next if message.author.bot?
   next unless message.content.start_with?("dist.beta.DocTest")
-  unless BotPerms::BETATESTERS.include?(message.author.id)
+  unless BotPerms::BETATESTERS.include?(message.author.id) or GuildLevels::BETAGUILDS.include?(message.guild.id)
      message.reply embed: Discorb::Embed.new("Error", "you are not a `BotPerms::BETATESTER`. if you know Ruby and Mongo, contact #{DistoolsStr::BCP} in order to become one of the `BotPerms::BETATESTERS`.", color: Discorb::Color.from_rgb(201, 0, 0))
      next
   end
@@ -248,7 +274,7 @@ end
 #end
 
 client.slash("whoami", "information about you") do |interaction|
-  interaction.post(embed: Discorb::Embed.new("User information for #{interaction.user.name}", "User ID: #{interaction.user.id}\nSignup time: #{interaction.user.created_at}\nUser flags: #{interaction.user.flag}", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+  interaction.post(embed: Discorb::Embed.new("User information for #{interaction.user.name}", "User ID: #{interaction.user.id}\nSignup time: #{interaction.user.created_at}\nUser flags: #{interaction.user.flag} - [calculate here](https://flags.lewisakura.moe)", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 end
 
 # Gonna see if this works. Ctrl C does not seem like a good way to end a bot so I
@@ -266,6 +292,8 @@ end
 client.slash("am-i-a-dev", "tells you whether the bot thinks you are a dev or not") do |interaction|
  if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
 	interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+ elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
  else
  if BotPerms::BOTOWNERS.include?(interaction.user.id) === true
 	if BotPerms::EXECUSERS.include?(interaction.user.id) === true
@@ -279,19 +307,10 @@ client.slash("am-i-a-dev", "tells you whether the bot thinks you are a dev or no
 end
 end
 
-client.slash("why-god-why", "[owner only] tests ABSOLUTELY EVERYTHING. could serve as a template command") do |interaction|
-  if BotPerms::BOTOWNERS.include?(interaction.user.id) === true
-	puts "Oh dear god it's a stress test. We're all gonna die."
-	interaction.post("*insert :hollow: here*", embed: Discorb::Embed.new("embed test :hollow:", "This is only a test"), ephemeral: false)
-  else
-	interaction.post("no", ephemeral: false)
-  end
-end
-
 client.slash("rift", "Open a rift to another server distools is in via your DMs [Owner only]", {
 	"server" => {
 	 required: true,
-	 type: :string,
+	 type: :int,
 	 description: "The server to open a rift to"
 	}
 }) do |interaction, server|
@@ -310,6 +329,8 @@ client.slash("say", "Get me to say something", {
 }) do |interaction, text|
 if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
   interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 else
  if text.include?("@here")
    interaction.post("Hey wait a minute thats not nice of you", ephemeral: false)
@@ -340,6 +361,8 @@ client.slash("timeout", "Timeout someone", {
 }) do |interaction, target, time, reason|
 if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
   interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 elsif interaction.user.permissions.mute_members? === true
  if reason
   timetoTO = Time.new + time.to_i
@@ -362,6 +385,8 @@ client.slash("tsay", "Say something without attributing it to you. For...reasons
 }) do |interaction, text|
 if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
   interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
 else
 if BotPerms::TUSERS.include?(interaction.user.id) === true
  if text.include?("@here")
@@ -390,15 +415,45 @@ else
 end
 end
 
-client.slash("brainpower", "LET THE BASS KICK") do |interaction|
-interaction.post("O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A- JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA")
-end
-
-
 client.once :standby do
   puts "ight its go time!! logged in as #{client.user}"  # Prints username of logged in user
   client.update_presence(status: :dnd)
   puts "guilds: Dictionary: #{client.guilds}"
+end
+
+# I pray this works.
+client.slash("timeout", "Timeout someone", {
+	"target" => {
+	 required: true,
+	 type: :user,
+	 description: "The person to timeout."
+	},
+        "time" => {
+         required: true,
+         type: :int,
+         description: "The amount of time (IN SECONDS) to time them out for"
+        },
+	"reason" => {
+	 required: false,
+	 type: :string,
+	 description: "Why are you doing this to them"
+	}
+}) do |interaction, target, reason|
+if BotPerms::BLACKLISTED.include?(interaction.user.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nYou have been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+elsif BotPerms::BLACKLISTEDGUILDS.include?(interaction.guild.id) === true
+  interaction.post(embed: Discorb::Embed.new("A critical exception has occurred.", "The command returned the following error:\nThis guild has been blacklisted from distools. Please contact #{DistoolsStr::BCP} if you believe this was in error.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+elsif interaction.user.permissions.ban_members? === true
+ if reason
+  target.ban(reason: "Banned by #{interaction.user.to_s} for committing the sin of #{reason.to_s}")
+  interaction.post(embed: Discorb::Embed.new("someones been bad :(", "I successfully **BANNED** out #{target.to_s} for committing the sin of \"#{reason.to_s}\"", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: true)
+ else
+  target.ban(reason: "Banned by #{interaction.user.to_s}, but they didn't say why")
+  interaction.post(embed: Discorb::Embed.new("someones been bad :(", "I just **BANNED** out #{target.to_s} for absolutely no reason.", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: true)
+ end
+else
+  interaction.post(embed: Discorb::Embed.new("Access denied.", "You do not have the permissions required by this command (**Ban Members**), please don't try to do that", color: Discorb::Color.from_rgb(201, 0, 0)), ephemeral: false)
+end
 end
 
 client.run ENV["TOKEN"]  # Starts client